+++ app/src/test/java/com/maximo/douglas/githubconsumer/components/gitpullrequestcard/context/GitPullRequestCardTest.kt
@@ -2,9 +2,9 @@ package com.maximo.douglas.githubconsumer.components.gitpullrequestcard.context
 
 import android.app.Activity
 import android.view.View
+import com.maximo.douglas.domain.entities.gitpullrequest.GitPullRequest
+import com.maximo.douglas.domain.entities.user.GitUser
 import com.maximo.douglas.githubconsumer.components.utils.BaseComponentTest
-import com.maximo.douglas.githubconsumer.domain.entities.gitpullrequest.GitPullRequest
-import com.maximo.douglas.githubconsumer.domain.entities.user.GitUser
 import com.maximo.douglas.githubconsumer.ui.components.pullrequestcard.GitPullRequestCardComponent
 import com.maximo.douglas.githubconsumer.utils.StringUtils.compactStringWithDots
 import kotlinx.android.synthetic.main.git_pull_request_card_component.view.*
