+++ app/src/test/java/com/maximo/douglas/githubconsumer/components/gitrepositorycard/context/GitRepositoryCardTest.kt
@@ -2,9 +2,9 @@ package com.maximo.douglas.githubconsumer.components.gitrepositorycard.context
 
 import android.app.Activity
 import android.view.View
+import com.maximo.douglas.domain.entities.gitrepository.GitRepository
+import com.maximo.douglas.domain.entities.user.GitUser
 import com.maximo.douglas.githubconsumer.components.utils.BaseComponentTest
-import com.maximo.douglas.githubconsumer.domain.entities.gitrepository.GitRepository
-import com.maximo.douglas.githubconsumer.domain.entities.user.GitUser
 import com.maximo.douglas.githubconsumer.ui.components.repositorycard.GitRepositoryCardComponent
 import com.maximo.douglas.githubconsumer.utils.StringUtils.compactStringWithDots
 import kotlinx.android.synthetic.main.git_repository_card_component.view.*
