+++ app/src/test/java/com/maximo/douglas/githubconsumer/viewmodels/gitpullrequest/context/SelectGitPullRequestViewModelTest.kt
@@ -2,7 +2,7 @@ package com.maximo.douglas.githubconsumer.viewmodels.gitpullrequest.context
 
 import androidx.arch.core.executor.testing.InstantTaskExecutorRule
 import androidx.lifecycle.Observer
-import com.maximo.douglas.githubconsumer.domain.entities.gitpullrequest.GitPullRequest
+import com.maximo.douglas.domain.entities.gitpullrequest.GitPullRequest
 import com.maximo.douglas.githubconsumer.service.gitpullrequest.GitPullRequestRemoteDataSource
 import com.maximo.douglas.githubconsumer.ui.gitpullrequests.SelectGitPullRequestViewModel
 import com.maximo.douglas.githubconsumer.ui.gitpullrequests.SelectGitPullRequestViewModel.SelectGitPullRequestViewModelState
