+++ domain/build.gradle
@@ -1,5 +1,5 @@
 plugins {
-    id 'com.android.application'
+    id 'com.android.library'
     id 'kotlin-android'
     id 'kotlin-android-extensions'
     id 'kotlin-kapt'
@@ -10,7 +10,6 @@ android {
     buildToolsVersion "28.0.3"
 
     defaultConfig {
-        applicationId "com.maximo.douglas.domain"
         minSdkVersion 24
         targetSdkVersion 28
         versionCode 1
